First implementation of Unit tests
 testFindCourseByValidId
Purpose: Validate the functionality of the method when provided with a valid id.

testDeleteCourseByValidId
Purpose: Validate the functionality of the method when provided with a valid id.

1.0.1 - testDeleteCourseByInvalidId
Purpose: Validate the functionality of the method when provided with an invalid id.

testFindCourseByInvalidId
Purpose: Validate the functionality of the method when provided with an invalid id.

testCreateValidCourse()
Purpose: Validate the functionality of the method when creation is successful

Modified EventService class for exception handling